+++ examples/cfd/external_aerodynamics/domino/src/deprecated/train_sharded.py
@@ -79,7 +79,7 @@ from physicsnemo.distributed import DistributedManager
 from physicsnemo.launch.utils import load_checkpoint, save_checkpoint
 from physicsnemo.launch.logging import PythonLogger, RankZeroLoggingWrapper
 
-from physicsnemo.datapipes.cae.domino_datapipe import (
+from physicsnemo.datapipes.cae.domino_datapipe2 import (
     compute_scaling_factors,
     create_domino_dataset,
 )