+++ physicsnemo/utils/neighbors/radius_search/_torch_impl.py
@@ -56,8 +56,6 @@ def radius_search_impl(
             dists = torch.empty((0,), device=dists.device, dtype=dists.dtype)
 
     else:
-        print(f"dists shape: {dists.shape}")
-
         # Take the max_points lowest distances for each query
         closest_points = torch.topk(
             dists, k=min(max_points, dists.shape[0]), dim=0, largest=False